moved query string logic from the form to the Search component
