+++ components/Search/forms/SearchCasesForm.spec.jsx
@@ -30,33 +30,16 @@ describe(`SearchCasesForm`, () => {
     await act(async () => {
       fireEvent.submit(getByRole('form'));
     });
-    expect(props.onFormSubmit).toHaveBeenCalledWith({
-      first_name: 'foo',
-      last_name: '',
-      case_note_type: '',
-      exact_name_match: false,
-      worker_email: '',
-    });
-  });
-
-  it('should pass the user email as worker_email', async () => {
-    const { getByRole, getByLabelText } = render(
-      <SearchCasesForm {...props} user={{ email: 'foo@bar.com' }} />
+    expect(props.onFormSubmit).toHaveBeenCalledWith(
+      {
+        first_name: 'foo',
+        last_name: '',
+        case_note_type: '',
+        exact_name_match: false,
+        my_notes_only: false,
+      },
+      expect.anything()
     );
-    const my_notes_onlyCheckbox = getByLabelText(
-      "Only include notes I've created"
-    );
-    fireEvent.click(my_notes_onlyCheckbox);
-    await act(async () => {
-      fireEvent.submit(getByRole('form'));
-    });
-    expect(props.onFormSubmit).toHaveBeenCalledWith({
-      first_name: '',
-      last_name: '',
-      case_note_type: '',
-      exact_name_match: false,
-      worker_email: 'foo@bar.com',
-    });
   });
 
   it('should initialise the form with the passed query', async () => {
@@ -66,12 +49,15 @@ describe(`SearchCasesForm`, () => {
     await act(async () => {
       fireEvent.submit(getByRole('form'));
     });
-    expect(props.onFormSubmit).toHaveBeenCalledWith({
-      first_name: 'bar',
-      last_name: '',
-      case_note_type: '',
-      exact_name_match: false,
-      worker_email: '',
-    });
+    expect(props.onFormSubmit).toHaveBeenCalledWith(
+      {
+        first_name: 'bar',
+        last_name: '',
+        case_note_type: '',
+        exact_name_match: false,
+        my_notes_only: false,
+      },
+      expect.anything()
+    );
   });
 });
