+++ components/Search/forms/SearchResidentsForm.spec.jsx
@@ -25,12 +25,15 @@ describe(`SearchResidentsForm`, () => {
     await act(async () => {
       fireEvent.submit(getByRole('form'));
     });
-    expect(props.onFormSubmit).toHaveBeenCalledWith({
-      first_name: 'foo',
-      last_name: '',
-      mosaic_id: '',
-      postcode: '',
-    });
+    expect(props.onFormSubmit).toHaveBeenCalledWith(
+      {
+        first_name: 'foo',
+        last_name: '',
+        mosaic_id: '',
+        postcode: '',
+      },
+      expect.anything()
+    );
   });
 
   it('should initialise the form with the passed query', async () => {
@@ -40,11 +43,14 @@ describe(`SearchResidentsForm`, () => {
     await act(async () => {
       fireEvent.submit(getByRole('form'));
     });
-    expect(props.onFormSubmit).toHaveBeenCalledWith({
-      first_name: 'bar',
-      last_name: '',
-      mosaic_id: '',
-      postcode: '',
-    });
+    expect(props.onFormSubmit).toHaveBeenCalledWith(
+      {
+        first_name: 'bar',
+        last_name: '',
+        mosaic_id: '',
+        postcode: '',
+      },
+      expect.anything()
+    );
   });
 });
